Dedicated skims

+++ SkimTuples/mergeSkimmedFiles.py
@@ -12,13 +12,15 @@ submission_parser = argParser.add_argument_group('submission', 'Arguments for su
 submission_parser.add_argument('--year',     action='store',      default=None,   help='Select year', choices=['2016', '2017', '2018'], required = True)
 submission_parser.add_argument('--overwrite', action='store_true', default=False,                help='overwrite if valid output file already exists')
 submission_parser.add_argument('--customList',  action='store',      default=None,               help='Name of a custom sample list. Otherwise it will use the appropriate noskim file.')
-submission_parser.add_argument('--skimSelection',  action='store',      default=None,               help='Name of the selection.')
+submission_parser.add_argument('--skimSelection',  action='store',      default='default',               help='Name of the selection.')
 submission_parser.add_argument('--skimName',  action='store',      default='Reco',               help='Name of the skim.')
 submission_parser.add_argument('--batchSystem', action='store',         default='HTCondor',  help='choose batchsystem', choices=['local', 'HTCondor', 'Cream02'])
 submission_parser.add_argument('--subJob',   action='store',      default=None,   help='The number of the subjob for this sample')
 submission_parser.add_argument('--isChild',  action='store_true', default=False,  help='mark as subjob, will never submit subjobs by itself')
 submission_parser.add_argument('--dryRun',   action='store_true', default=False,  help='do not launch subjobs, only show them')
-
+submission_parser.add_argument('--isTest',   action='store_true', default=False,  help='Run a small test')
+submission_parser.add_argument('--region',   action='store', default=None,  help='Choose the selection region', 
+    choices=['baseline', 'highMassSR', 'lowMassSR', 'ZZCR', 'WZCR', 'ConversionCR'])
 
 args = argParser.parse_args()
 
@@ -27,7 +29,7 @@ if not args.isChild:
     if abort_script == 'n' or abort_script == 'N':
         exit(0)
 
-if args.batchSystem == 'HTCondor' and not args.isChild:
+if args.batchSystem == 'HTCondor' and not args.isChild and not args.isTest and args.region is None:
     jobs = [[0]]
     submitJobs(__file__, ('subJob'), jobs, argParser, jobLabel = 'mergeSkim')
     exit(0)
@@ -36,8 +38,12 @@ from HNL.Tools.helpers import fileSize, makeDirIfNeeded, isValidRootFile
 
 def mergeSmallFile(merge_file):
     path, name = merge_file.rsplit('/', 1)
-    os.system('hadd -f -v '+path+'/'+name.split('_', 1)[1]+'.root '+merge_file+'/*root')
-    os.system('rm -r '+ merge_file)
+    if not args.isTest:
+        os.system('hadd -f -v '+path+'/'+name.split('_', 1)[1]+'.root '+merge_file+'/*root')
+        os.system('rm -r -f '+ merge_file)
+    else:
+        makeDirIfNeeded(path+'/testArea/'+name.split('_', 1)[1]+'.root')
+        os.system('hadd -f -v '+path+'/testArea/'+name.split('_', 1)[1]+'.root '+merge_file+'/*root')
 
 def mergeLargeFile(merge_file):
     path, name = merge_file.rsplit('/', 1)
@@ -45,10 +51,13 @@ def mergeLargeFile(merge_file):
     split_list = []
     tmp_list = []
 
-    makeDirIfNeeded(os.path.join(path, name, 'tmp_batches', 'x'))
+    if args.isTest:
+        makeDirIfNeeded(os.path.join(path, 'testArea', name, 'tmp_batches', 'x'))
+    else:
+        makeDirIfNeeded(os.path.join(path, name, 'tmp_batches', 'x'))
 
     for i, file_list in enumerate(sub_files):
-        if i != 0 and (i % 20 == 0 or i == len(merge_files)-1):
+        if i != 0 and (i % 20 == 0 or i == len(sub_files)-1):
             tmp_list.append(file_list)
             split_list.append([x for x in tmp_list])
             tmp_list = []
@@ -56,22 +65,29 @@ def mergeLargeFile(merge_file):
             tmp_list.append(file_list)
 
     for j, file_list in enumerate(split_list):
-        os.system('hadd -f -v '+path+'/'+name+ '/tmp_batches/batch_'+str(j)+ '.root '+' '.join(file_list))
-        for f in file_list:
-            os.system('rm '+f)
+        if not args.isTest:
+            os.system('hadd -f -v '+path+'/'+name+ '/tmp_batches/batch_'+str(j)+ '.root '+' '.join(file_list))
+            for f in file_list:
+                os.system('rm '+f)
+        else:
+            os.system('hadd -f -v '+path+'/testArea/'+name+ '/tmp_batches/batch_'+str(j)+ '.root '+' '.join(file_list))
 
-    os.system('hadd -f -v '+path+'/'+name.split('_', 1)[1]+'.root '+path+'/'+name+ '/tmp_batches/*root')
-    os.system('rm -rf '+path+'/'+name)
+    if args.isTest:
+        os.system('hadd -f -v '+path+'/testArea/'+name.split('_', 1)[1]+'.root '+path+'/testArea/'+name+ '/tmp_batches/*root')
+    else:
+        os.system('hadd -f -v '+path+'/'+name.split('_', 1)[1]+'.root '+path+'/'+name+ '/tmp_batches/*root')
+        os.system('rm -rf '+path+'/'+name)
 
-pnfs_base = os.path.join('/pnfs/iihe/cms/store/user', os.path.expandvars('$USER'), 'skimmedTuples/HNL', args.skimSelection, args.year, args.skimName)
-pnfs_backup_base = os.path.join('/pnfs/iihe/cms/store/user', os.path.expandvars('$USER'), 'skimmedTuples/HNL/Backup', args.skimSelection, args.year, args.skimName)
+skim_selection_string = args.skimSelection if args.region is None else args.skimSelection+'/'+args.region 
+pnfs_base = os.path.join('/pnfs/iihe/cms/store/user', os.path.expandvars('$USER'), 'skimmedTuples/HNL', skim_selection_string, args.year, args.skimName)
+pnfs_backup_base = os.path.join('/pnfs/iihe/cms/store/user', os.path.expandvars('$USER'), 'skimmedTuples/HNL/Backup', skim_selection_string, args.year, args.skimName)
 if not args.batchSystem == 'HTCondor':
     pnfs_base = 'srm://maite.iihe.ac.be:8443'+pnfs_base
     pnfs_backup_base = 'srm://maite.iihe.ac.be:8443'+pnfs_backup_base
 
 #Merges subfiles if needed
 if args.batchSystem != 'HTCondor':
-    merge_files = glob.glob('/storage_mnt/storage/user/'+os.path.expandvars('$USER')+'/public/ntuples/HNL/'+args.skimSelection+'/'+args.year+'/'+args.skimName+'/tmp*')
+    merge_files = glob.glob('/storage_mnt/storage/user/'+os.path.expandvars('$USER')+'/public/ntuples/HNL/'+skim_selection_string+'/'+args.year+'/'+args.skimName+'/tmp*')
 else:
     merge_files = glob.glob(pnfs_base+'/tmp*')
 merge_files = sorted(merge_files)
@@ -88,13 +104,14 @@ for mf in merge_files:
     path, name = mf.rsplit('/', 1)
     new_name = name.split('_', 1)[1]+'.root'
 
-    if args.batchSystem != 'HTCondor':
-        os.system('gfal-rm '+pnfs_backup_base+'/'+new_name)
-        os.system('gfal-copy '+pnfs_base+'/'+new_name+' '+pnfs_backup_base+'/'+new_name)
-        os.system('gfal-rm '+pnfs_base+'/'+new_name)
-    else:
-        if isValidRootFile(pnfs_base+'/'+new_name):
-            os.system('scp '+pnfs_base+'/'+new_name + ' '+ pnfs_backup_base+'/'+new_name)
+    if not args.isTest:
+        if args.batchSystem != 'HTCondor':
+            os.system('gfal-rm '+pnfs_backup_base+'/'+new_name)
+            os.system('gfal-copy '+pnfs_base+'/'+new_name+' '+pnfs_backup_base+'/'+new_name)
+            os.system('gfal-rm '+pnfs_base+'/'+new_name)
+        else:
+            if isValidRootFile(pnfs_base+'/'+new_name):
+                os.system('scp '+pnfs_base+'/'+new_name + ' '+ pnfs_backup_base+'/'+new_name)
 
     sub_files = glob.glob(mf+'/*')
     tot_size = 0
@@ -105,8 +122,10 @@ for mf in merge_files:
         mergeLargeFile(mf)
     else:
         mergeSmallFile(mf)
+        
+    # mergeLargeFile(mf)
 
-    if args.batchSystem != 'HTCondor':
+    if args.batchSystem != 'HTCondor' and not args.isTest:
         #Move to pnfs once it is merged locally
         os.system('gfal-copy file://'+path+'/'+new_name+' '+pnfs_base+'/'+new_name)
         os.system('rm '+path+'/'+new_name)
